Models and migrations moved to database folder
