+++ BookShop/.eslintrc.json
@@ -1,20 +1,24 @@
 {
+    "extends": "standard",
+    "plugins": [
+        "standard",
+        "promise",
+        "chai-friendly"
+    ],
     "env": {
-        "browser": true,
-        "commonjs": true,
-        "es6": true
-    },
-    "extends": "eslint:recommended",
-    "globals": {
-        "Atomics": "readonly",
-        "SharedArrayBuffer": "readonly"
-    },
-    "parserOptions": {
-        "ecmaVersion": 2018
-    },
+         "node": true,
+         "mocha": true
+     },
     "rules": {
-		"semi": ["error", "always"],
-        "quotes": ["error", "double"],
-		"no-var": 0
+        "eol-last": "off",
+        "space-before-function-paren": "off",
+        "indent": ["error", 4],
+        "quotes": ["error", "single", { "avoidEscape": true }],
+        "semi": ["error", "always",{ "omitLastInOneLineBlock": true}],
+        "one-var": ["off"],
+        "space-before-blocks": ["off"],
+        "camelcase": ["warn"],
+        "no-unused-expressions": 0,
+        "chai-friendly/no-unused-expressions": 2
     }
-}
+  }
