+++ BookShop/.sequelizerc
@@ -2,7 +2,7 @@
 
 module.exports = {
   "config": path.resolve('./server/config', 'config.json'),
-  "models-path": path.resolve('./server/models'),
+  "models-path": path.resolve('./server/database/models'),
   "seeders-path": path.resolve('./server/seeders'),
-  "migrations-path": path.resolve('./server/migrations')
+  "migrations-path": path.resolve('./server/database/migrations')
 };
