+++ BookShop/BookShop.njsproj
@@ -50,20 +50,20 @@
     <Compile Include="server\middleware\index.js">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="server\migrations\20190221205315-create-page.js" />
-    <Compile Include="server\migrations\20190221212005-create-book.js" />
-    <Compile Include="server\migrations\20190221212637-create-author.js" />
-    <Compile Include="server\migrations\20190221212946-create-book-author.js" />
-    <Compile Include="server\migrations\20190224183937-create-user.js" />
+    <Compile Include="server\database\migrations\20190221205315-create-page.js" />
+    <Compile Include="server\database\migrations\20190221212005-create-book.js" />
+    <Compile Include="server\database\migrations\20190221212637-create-author.js" />
+    <Compile Include="server\database\migrations\20190221212946-create-book-author.js" />
+    <Compile Include="server\database\migrations\20190224183937-create-user.js" />
     <Compile Include="server\controllers\auth.js">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="server\models\author.js" />
-    <Compile Include="server\models\book.js" />
-    <Compile Include="server\models\bookauthor.js" />
-    <Compile Include="server\models\index.js" />
-    <Compile Include="server\models\page.js" />
-    <Compile Include="server\models\user.js" />
+    <Compile Include="server\database\models\author.js" />
+    <Compile Include="server\database\models\book.js" />
+    <Compile Include="server\database\models\bookauthor.js" />
+    <Compile Include="server\database\models\index.js" />
+    <Compile Include="server\database\models\page.js" />
+    <Compile Include="server\database\models\user.js" />
     <Compile Include="server\routes\auth.js">
       <SubType>Code</SubType>
     </Compile>
@@ -98,8 +98,8 @@
   <ItemGroup>
     <Folder Include="server\" />
     <Folder Include="server\config\" />
-    <Folder Include="server\migrations\" />
-    <Folder Include="server\models\" />
+    <Folder Include="server\database\migrations\" />
+    <Folder Include="server\database\models\" />
     <Folder Include="server\controllers\" />
     <Folder Include="server\middleware\" />
     <Folder Include="server\database\" />
