+++ BookShop/package.json
@@ -30,13 +30,13 @@
     "swagger-ui-express": "^4.0.2"
   },
   "devDependencies": {
-    "eslint": "^5.14.1"
-	"bcrypt-nodejs": "0.0.3",
+    "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.15.0",
     "chai": "^4.2.0",
     "chai-http": "^4.2.1",
     "cookie-parser": "^1.4.0",
     "debug": "^2.2.0",
+    "eslint": "^5.14.1",
     "express": "^4.14.0",
     "express-session": "^1.15.6",
     "g": "^2.0.1",
