+++ BookShop/server/controllers/auth.js
@@ -1,4 +1,4 @@
-﻿const User = require("../models").User;
+﻿const User = require("../database/models").User;
 const jwt = require("jsonwebtoken");
 const config = require("../config");
